Move start/stop commands code to separate modules
