+++ cli/pcluster/commands.py
@@ -266,53 +266,6 @@ def _is_ganglia_enabled(parameters):
     return is_ganglia_enabled
 
 
-def start(args):
-    """Restore ASG limits or awsbatch CE to min/max/desired."""
-    stack_name = utils.get_stack_name(args.cluster_name)
-    pcluster_config = PclusterConfig(config_file=args.config_file, cluster_name=args.cluster_name)
-    cluster_section = pcluster_config.get_section("cluster")
-
-    if cluster_section.get_param_value("scheduler") == "awsbatch":
-        LOGGER.info("Enabling AWS Batch compute environment : %s", args.cluster_name)
-        max_vcpus = cluster_section.get_param_value("max_vcpus")
-        desired_vcpus = cluster_section.get_param_value("desired_vcpus")
-        min_vcpus = cluster_section.get_param_value("min_vcpus")
-        ce_name = utils.get_batch_ce(stack_name)
-        _start_batch_ce(ce_name=ce_name, min_vcpus=min_vcpus, desired_vcpus=desired_vcpus, max_vcpus=max_vcpus)
-    elif cluster_section.get_param_value("scheduler") == "slurm":
-        # TODO: to be implemented
-        raise NotImplementedError()
-    else:
-        LOGGER.info("Starting compute fleet: %s", args.cluster_name)
-        max_queue_size = cluster_section.get_param_value("max_queue_size")
-        min_desired_size = (
-            cluster_section.get_param_value("initial_queue_size")
-            if cluster_section.get_param_value("maintain_initial_size")
-            else 0
-        )
-        asg_name = utils.get_asg_name(stack_name)
-        utils.set_asg_limits(asg_name=asg_name, min=min_desired_size, max=max_queue_size, desired=min_desired_size)
-
-
-def stop(args):
-    """Set ASG limits or awsbatch ce to min/max/desired = 0/0/0."""
-    stack_name = utils.get_stack_name(args.cluster_name)
-    pcluster_config = PclusterConfig(config_file=args.config_file, cluster_name=args.cluster_name)
-    cluster_section = pcluster_config.get_section("cluster")
-
-    if cluster_section.get_param_value("scheduler") == "awsbatch":
-        LOGGER.info("Disabling AWS Batch compute environment : %s", args.cluster_name)
-        ce_name = utils.get_batch_ce(stack_name)
-        _stop_batch_ce(ce_name=ce_name)
-    elif cluster_section.get_param_value("scheduler") == "slurm":
-        # TODO: to be implemented
-        raise NotImplementedError()
-    else:
-        LOGGER.info("Stopping compute fleet: %s", args.cluster_name)
-        asg_name = utils.get_asg_name(stack_name)
-        utils.set_asg_limits(asg_name=asg_name, min=0, max=0, desired=0)
-
-
 def _get_pcluster_version_from_stack(stack):
     """
     Get the version of the stack if tagged.
@@ -410,29 +363,9 @@ def _get_compute_instances(stack):
     return map(lambda i: ("ComputeFleet", i.get("InstanceId")), instances)
 
 
-def _start_batch_ce(ce_name, min_vcpus, desired_vcpus, max_vcpus):
-    try:
-        boto3.client("batch").update_compute_environment(
-            computeEnvironment=ce_name,
-            state="ENABLED",
-            computeResources={
-                "minvCpus": int(min_vcpus),
-                "maxvCpus": int(max_vcpus),
-                "desiredvCpus": int(desired_vcpus),
-            },
-        )
-    except ClientError as e:
-        LOGGER.critical(e.response.get("Error").get("Message"))
-        sys.exit(1)
-
-
-def _stop_batch_ce(ce_name):
-    boto3.client("batch").update_compute_environment(computeEnvironment=ce_name, state="DISABLED")
-
-
 def instances(args):
     stack_name = utils.get_stack_name(args.cluster_name)
-    pcluster_config = PclusterConfig(config_file=args.config_file, cluster_name=args.cluster_name)
+    pcluster_config = PclusterConfig(config_file=args.config_file, cluster_name=args.cluster_name, auto_refresh=False)
     cluster_section = pcluster_config.get_section("cluster")
 
     instances = []
